Added a conditional to check user type in UserProfile component
