+++ Tabloid/client/src/components/userProfile/UserProfileList.js
@@ -1,17 +1,15 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { UserProfileContext } from '../../providers/UserProfileProvider';
-import { useHistory } from 'react-router-dom';
 import UserDeactivationModal from './UserDeactivationModal';
-import { Table, Button, Modal, ModalBody, ModalHeader } from "reactstrap";
+import { Table, Modal, ModalBody, ModalHeader } from "reactstrap";
 import { EditUserProfileForm } from "../EditUserProfileForm";
 import { UserTypeContext } from '../../providers/UserTypeProvider';
-import AdminUserProfile from './AdminUserProfile';
-import AuthorUserProfile from './AuthorUserProfile';
+import UserProfile from './UserProfile';
+
 
 export default () => {
     const { userProfiles, getAllUserProfiles } = useContext(UserProfileContext)
     const { getAllUserTypes } = useContext(UserTypeContext)
-    const currentUser = JSON.parse(sessionStorage.getItem("userProfile"))
     const activeUsers = userProfiles.filter(up => up.isActive === true)
     const [modal, setModal] = useState(false)
     const toggleModal = () => setModal(!modal)
@@ -19,7 +17,6 @@ export default () => {
     const [selectedUser, setSelectedUser] = useState({});
     const [editModal, setEditModal] = useState(false);
     const toggleEditModal = () => setEditModal(!editModal);
-    const history = useHistory()
 
     useEffect(() => {
         getAllUserProfiles()
@@ -41,71 +38,15 @@ export default () => {
                 </thead>
                 <tbody>
                     {
-                        (currentUser.userType.name === "Admin")
-                            ?
-                            (
-                                activeUsers.map(profile => {
-                                    return <AdminUserProfile
-                                        user={profile}
-                                        setClickedUser={setClickedUser}
-                                        toggleModal={toggleModal}
-                                        setSelectedUser={setSelectedUser}
-                                        toggleEditModal={toggleEditModal}
-                                    />
-                                })
-                            )
-                            :
-                            (
-                                activeUsers.map(profile => {
-                                    return <AuthorUserProfile user={profile} />
-                                })
-                            )
-
-                        // (activeUsers.map(profile => {
-
-                        //     return (
-                        //         <tr key={profile.id}>
-                        //             <td style={{ cursor: "pointer" }} onClick={() => history.push(`/userProfile/${profile.id}`)}>
-                        //                 {profile.fullName}
-                        //             </td>
-                        //             <td>{profile.displayName}</td>
-                        //             <td>{profile.userType.name}</td>
-                        //             <td style={{ color: "blue", cursor: "pointer" }}
-                        //                 onClick={() => {
-                        //                     setClickedUser(profile)
-                        //                     toggleModal()
-                        //                 }}
-                        //             >
-                        //                 Deactivate
-                        //             </td>
-                        //             <td>
-                        //                 <Button
-                        //                     onClick={() => {
-                        //                         setSelectedUser(profile);
-                        //                         toggleEditModal();
-                        //                     }}
-                        //                 >
-                        //                     Edit
-                        //                 </Button>
-                        //             </td>
-                        //         </tr>
-                        //     )
-                        // }))
-                        // :
-                        // (
-                        //     (activeUsers.map(profile => {
-
-                        //         return (
-                        //             <tr key={profile.id}>
-                        //                 <td style={{ cursor: "pointer" }} onClick={() => history.push(`/userProfile/${profile.id}`)}>
-                        //                     {profile.fullName}
-                        //                 </td>
-                        //                 <td>{profile.displayName}</td>
-                        //                 <td>{profile.userType.name}</td>
-                        //             </tr>
-                        //         )
-                        //     }))
-                        // )
+                        activeUsers.map(profile => {
+                            return <UserProfile
+                                user={profile}
+                                setClickedUser={setClickedUser}
+                                toggleModal={toggleModal}
+                                setSelectedUser={setSelectedUser}
+                                toggleEditModal={toggleEditModal}
+                            />
+                        })
                     }
                 </tbody>
             </Table>
